chore: moves AnalyticsHandler to app folder
