+++ packages/core/app/sdk/analytics/hooks/useViewItemListEvent.ts
