+++ packages/core/app/sdk/analytics/platform/vtex/index.ts
