+++ packages/core/src/sdk/cart/useBuyButton.ts
@@ -1,12 +1,12 @@
+import type { AddToCartEvent, CurrencyCode } from '@faststore/sdk'
 import { sendAnalyticsEvent } from '@faststore/sdk'
 import { useCallback } from 'react'
-import type { CurrencyCode, AddToCartEvent } from '@faststore/sdk'
 
-import type { AnalyticsItem } from 'src/sdk/analytics/types'
+import type { AnalyticsItem } from 'app/sdk/analytics/types'
 import type { CartItem } from 'src/sdk/cart'
 
-import { useSession } from '../session'
 import { useUI } from '@faststore/ui'
+import { useSession } from '../session'
 import { cartStore } from './index'
 
 export const useBuyButton = (item: CartItem | null) => {
